Extract exclude_patterns module from controller.py (#394 Phase 1A)
Move 4 pure functions (_matches_exclude, _filter_children,
parse_exclude_patterns, filter_excluded_files) from controller.py to
controller/exclude_patterns.py.

- Remove Controller._apply_exclude_patterns thin wrapper; call site
  now calls filter_excluded_files directly
- Update __init__.py to re-export from new module (public API unchanged)
- Remove unused fnmatch and SystemFile imports from controller.py

controller.py: 1433 → 1339 lines (-94)
All 28 existing exclude pattern tests pass unchanged.
