+++ src/python/controller/controller.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import copy
-import fnmatch
 import logging
 import os
 from abc import ABC, abstractmethod
@@ -14,7 +13,7 @@ from threading import Lock
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from system import SystemFile
+    pass
 
 from common import AppError, AppOneShotProcess, Constants, Context, MultiprocessingLogger
 from lftp import Lftp, LftpError, LftpJobStatus
@@ -22,83 +21,14 @@ from model import IModelListener, Model, ModelDiff, ModelDiffUtil, ModelError, M
 
 from .controller_persist import ControllerPersist
 from .delete import DeleteLocalProcess, DeleteRemoteProcess
+
+# my libs
+from .exclude_patterns import filter_excluded_files, parse_exclude_patterns
 from .extract import ExtractProcess, ExtractRequest, ExtractStatus, ExtractStatusResult
 from .model_builder import ModelBuilder
 from .move import MoveProcess
-
-# my libs
 from .scan import ActiveScanner, LocalScanner, RemoteScanner, ScannerProcess, ScannerResult
 from .validate import ValidateProcess, ValidateRequest, ValidateStatusResult
-
-
-def _matches_exclude(name: str, is_dir: bool, patterns: list[tuple[str, bool]]) -> bool:
-    """Return True if *name* matches any of the exclude patterns (case-insensitive).
-
-    Each pattern is a (glob, dir_only) tuple.  When dir_only is True the
-    pattern only matches directories.
-    """
-    name_lower = name.lower()
-    return any(fnmatch.fnmatch(name_lower, p.lower()) and (not dir_only or is_dir) for p, dir_only in patterns)
-
-
-def _filter_children(file: SystemFile, patterns: list[tuple[str, bool]]) -> SystemFile:
-    """Return a copy of *file* with excluded children (and their subtrees) removed.
-
-    If a directory child matches a pattern the entire subtree is dropped.
-    Non-matching directory children are recursed into so their own children
-    are filtered as well.  Directory sizes are preserved from the original file.
-    """
-    from system import SystemFile  # avoid circular import at module level
-
-    kept_children: list[SystemFile] = []
-    for child in file.children:
-        if _matches_exclude(child.name, child.is_dir, patterns):
-            continue  # drop matched child (and its subtree)
-        if child.is_dir:
-            child = _filter_children(child, patterns)
-        kept_children.append(child)
-
-    filtered = SystemFile(
-        name=file.name,
-        size=file.size,
-        is_dir=file.is_dir,
-        time_created=file.timestamp_created,
-        time_modified=file.timestamp_modified,
-    )
-    for child in kept_children:
-        filtered.add_child(child)
-    return filtered
-
-
-def parse_exclude_patterns(exclude_patterns_str: str) -> list[str]:
-    """Parse a comma-separated exclude pattern string into a list of individual patterns.
-
-    Trailing slashes are preserved so callers can distinguish directory-only
-    patterns from file patterns when needed.
-    """
-    if not exclude_patterns_str or not exclude_patterns_str.strip():
-        return []
-    patterns: list[str] = []
-    for p in exclude_patterns_str.split(","):
-        p = p.strip()
-        if p:
-            patterns.append(p)
-    return patterns
-
-
-def filter_excluded_files(files: list[SystemFile], exclude_patterns_str: str) -> list[SystemFile]:
-    parsed = parse_exclude_patterns(exclude_patterns_str)
-    if not parsed:
-        return files
-    patterns = [(p.rstrip("/"), p.endswith("/")) for p in parsed]
-    result: list[SystemFile] = []
-    for f in files:
-        if _matches_exclude(f.name, f.is_dir, patterns):
-            continue
-        if f.is_dir:
-            f = _filter_children(f, patterns)
-        result.append(f)
-    return result
 
 
 class ControllerError(AppError):
@@ -1037,7 +967,9 @@ class Controller:
         pc.model_builder.set_auto_delete_remote(bool(self.__context.config.autoqueue.auto_delete_remote))
 
         if latest_remote_scan is not None:
-            remote_files = self._apply_exclude_patterns(latest_remote_scan.files)
+            remote_files = filter_excluded_files(
+                latest_remote_scan.files, self.__context.config.general.exclude_patterns
+            )
             pc.model_builder.set_remote_files(remote_files)
         if latest_local_scan is not None:
             pc.model_builder.set_local_files(latest_local_scan.files)
@@ -1405,7 +1337,3 @@ class Controller:
                 for pc in self.__pair_contexts:
                     pc.local_scan_process.force_scan()
         self.__active_move_processes = still_active_moves
-
-    def _apply_exclude_patterns(self, files: list[SystemFile]) -> list[SystemFile]:
-        raw = self.__context.config.general.exclude_patterns
-        return filter_excluded_files(files, raw)
